requirements.txt
django
django-rest-framework
django_filter==22.1
djangorestframework==3.14.0
